return float

+++ 0x01-python_async_function/2-measure_runtime.py
@@ -18,4 +18,4 @@ function should return a float.
     start = time.perf_counter()
     asyncio.run(wait_n(n, max_delay))
     elapsed = time.perf_counter() - start
-    return elapsed / n
+    return float(elapsed / n)
